+++ Changelog.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## 3.6.10 / 2021-05-20
+
+### Fixed:
+
+  * Better error handling for version check
+  * Fix broken `isTerminated()`
+  * Use github actions for creating automatic releases
+  * Run php unit tests in github actions
+
+
+## 3.6.9 / 2021-05-19
+
+### Fixed:
+
+  * Fix missing method in DryRunShellProvider, add test coverage
+
+## 3.6.8 / 2021-05-19
+
+### Fixed:
+
+* Terminate current shell, when new app gets deployed via k8s
+
 ## 3.6.7 / 2021-05-19
 
 ### Changed: